1) refactored the src/server folder structure 1.1) added static_content/ where all the index.html, css, etc goes 1.2) added static_content/client_dist/ where babel dumps the bundle.js file
this allows the babel to be run on src/server/main and all the js files underneath that folder structure when the server is more complex and uses TS / Flow / ES6, etc. It also separates the entry pt of the server from the static_content and web app files. Also, src/server/main can access src/global which is a good thing to share code between webapp and server